+++ packages/plugin-windows/src/windows-info-service.js
@@ -1,8 +1,11 @@
+import DetectEngine from 'appcd-detect';
 import gawk from 'gawk';
-import windowslib from 'windowslib';
+import version from './version';
+
+import * as windowslib from 'windowslib';
 
 import { DataServiceDispatcher } from 'appcd-dispatcher';
-import { get } from 'appcd-util';
+import { arrayify, get, mergeDeep } from 'appcd-util';
 
 /**
  * The Windows info service.
@@ -17,201 +20,103 @@ export default class WindowsInfoService extends DataServiceDispatcher {
 	 * @access public
 	 */
 	async activate(cfg) {
+		this.config = cfg;
+		if (cfg.windows) {
+			mergeDeep(windowslib.options, cfg.windows);
+		}
+
 		this.data = gawk({
-			windows: {},
-			visualstudio: {},
-			selectedVisualStudio: {}
+			sdks: {},
+			vs: {}
 		});
 
-		this.timers = {};
+		/**
+		 * A map of buckets to a list of active fs watch subscription ids.
+		 * @type {Object}
+		 */
+		this.subscriptions = {};
 
-		// wire up Visual Studio detection first so that we can use its result to know if we should query the other thing
-		await this.wireupDetection('visualstudio', get(cfg, 'windows.visualstudio.pollInterval') || 60000 * 10, () => this.detectVisualStudios());
-		await this.wireupDetection('windows',      get(cfg, 'windows.windowsSDK.pollInterval')   || 60000 / 2,  () => this.detectWindowsSDKs());
+		/**
+		 * ?
+		 */
+		// this.winregWatchHandles = {};
+
+		await Promise.all([
+			this.initSDKs() // ,
+			// this.initVisualStudios()
+		]);
 	}
 
 	/**
-	 * Stops all active timers.
+	 * Shutsdown the Windows info service.
 	 *
+	 * @returns {Promise}
 	 * @access public
 	 */
-	deactivate() {
-		for (const timer of Object.values(this.timers)) {
-			clearTimeout(timer);
+	async deactivate() {
+		// for (const handle of Object.values(this.winregWatchHandles)) {
+		// 	handle.stop();
+		// }
+
+		if (this.sdkDetectEngine) {
+			await this.sdkDetectEngine.stop();
+			this.sdkDetectEngine = null;
 		}
-		this.timers = {};
 	}
 
 	/**
-	 * Executes a detect function, then stores the result and schedules the next check.
+	 * Detect Windows SDKs.
 	 *
-	 * @param {String} type - The bucket name for the detected results.
-	 * @param {Number} interval - The amount of milliseconds until the next check.
-	 * @param {Function} callback - A function to call that performs the detection.
 	 * @returns {Promise}
 	 * @access private
 	 */
-	async wireupDetection(type, interval, callback) {
-		try {
-			const result = await callback();
-			if (result) {
-				console.log(`Updating data for ${type}`);
-				gawk.set(this.data[type], result);
-			}
-		} catch (err) {
-			console.log(err);
-		}
+	async initSDKs() {
+		const paths = [
+			...arrayify(get(this.config, 'windows.sdk.searchPaths'), true),
+			windowslib.sdk.defaultPath
+		];
 
-		this.timers[type] = setTimeout(() => {
-			this.wireupDetection(type, interval, callback);
-		}, interval);
-	}
-
-	/**
-	 * Checks if there are any Visual Studios installed.
-	 *
-	 * @returns {Boolean}
-	 * @access private
-	 */
-	haveVisualStudio() {
-		return Object.keys(this.data.visualstudio).length > 0;
-	}
-
-	/**
-	 * Detect Windows Phone devices.
-	 *
-	 * @returns {Promise<Array.<Object>>}
-	 * @access private
-	 */
-	detectDevices() {
-		return new Promise((resolve, reject) => {
-			if (!this.haveVisualStudio()) {
-				return resolve();
-			}
-
-			console.log('Detecting devices info');
-			windowslib.device.detect({ bypassCache: true }, (err, results) => {
-				if (err) {
-					reject(err);
-				} else {
-					const devices = results.devices;
-					let wpsdkIndex = -1;
-					let realDeviceIndex = -1;
-					for (let i = 0; i < devices.length; i++) {
-						const device = devices[i];
-						if (device.udid === 0 && device.wpsdk) {
-							wpsdkIndex = i;
-						} else if (device.udid !== 0 && !device.wpsdk) {
-							// now find with "real" device
-							realDeviceIndex = i;
-						}
-						if (wpsdkIndex !== -1 && realDeviceIndex !== -1) {
-							break;
-						}
-					}
-					if (wpsdkIndex !== -1 && realDeviceIndex !== -1) {
-						// set 'real' device wpsdk to the value we got from wptool binary
-						devices[realDeviceIndex].wpsdk = devices[wpsdkIndex].wpsdk;
-						// remove the wptool binary entry
-						devices.splice(wpsdkIndex, 1);
-					}
-					resolve(devices);
+		this.sdkDetectEngine = new DetectEngine({
+			checkDir(dir) {
+				try {
+					return windowslib.sdk.detectSDKs(dir);
+				} catch (e) {
+					// 'dir' is not an SDK
 				}
-			});
+			},
+			depth:               1,
+			multiple:            true,
+			name:                'windows:sdk',
+			paths,
+			recursive:           true,
+			recursiveWatchDepth: 3,
+			redetect:            true,
+			// registryKeys:        windowslib.sdk.registryKeys.map(key => ({
+			// 	key: new RegExp(key + )
+			// })),
+			watch:               true
 		});
-	}
 
-	/**
-	 * Detect Windows Phone emulators.
-	 *
-	 * @returns {Promise<Object>}
-	 * @access private
-	 */
-	detectEmulators() {
-		return new Promise((resolve, reject) => {
-			if (!this.haveVisualStudio()) {
-				return resolve();
+		// listen for sdk results
+		this.sdkDetectEngine.on('results', results => {
+			const sdks = {};
+			for (const sdk of results.sort((a, b) => version.compare(a.version, b.version))) {
+				sdks[sdk.version] = sdk;
 			}
-
-			console.log('Detecting emulator info');
-			windowslib.emulator.detect({ bypassCache: true }, (err, results) => {
-				if (err) {
-					reject(err);
-				} else {
-					resolve(results.emulators);
-				}
-			});
+			gawk.set(this.data.sdks, sdks);
 		});
+
+		await this.sdkDetectEngine.start();
 	}
 
 	/**
 	 * Detect Visual Studio installations.
 	 *
-	 * @returns {Promise<Object>}
+	 * @returns {Promise}
 	 * @access private
-	 */
-	detectVisualStudios() {
-		return new Promise((resolve, reject) => {
-			console.log('Detecting visualstudio info');
-			windowslib.visualstudio.detect({ bypassCache: true }, (err, results) => {
-				if (err) {
-					return reject(err);
-				}
-
-				let found = false;
-				if (results.visualstudio) {
-					for (const visualstudio of Object.keys(results.visualstudio)) {
-						if (results.visualstudio[visualstudio].selected) {
-							found = true;
-							gawk.set(this.data.selectedVisualStudio, results.visualstudio[visualstudio]);
-							break;
-						}
-					}
-				}
-				if (!found) {
-					this.data.selectedVisualStudio = null;
-				}
-
-				resolve(results.visualstudio);
-			});
-		});
+	 * /
+	initVisualStudios() {
+		//
 	}
-
-	/**
-	 * Detect Windows Store SDK information.
-	 *
-	 * @returns {Promise<Object>}
-	 * @access private
-	 */
-	detectWindowsSDKs() {
-		return new Promise((resolve, reject) => {
-			console.log('Detecting windows store info');
-			windowslib.winstore.detect({ bypassCache: true }, (err, results) => {
-				if (err) {
-					reject(err);
-				} else {
-					resolve(results.windows);
-				}
-			});
-		});
-	}
-
-	/**
-	 * Detect Windows Phone SDK information.
-	 *
-	 * @returns {Promise<Object>}
-	 * @access private
-	 */
-	detectWindowsPhone() {
-		return new Promise((resolve, reject) => {
-			console.log('Detecting windowsphone info');
-			windowslib.windowsphone.detect({ bypassCache: true }, (err, results) => {
-				if (err) {
-					reject(err);
-				} else {
-					resolve(results.windowsphone);
-				}
-			});
-		});
-	}
+	*/
 }